+++ app/javascript/packs/application.js
@@ -32,7 +32,8 @@ import { initChangeImage } from "./change_image";
 import { advancedSearch } from "./advanced_search";
 import {bubble} from './bubble_value_range';
 import { initUpdateNavbarOnScroll } from '../components/navbar';
-import { initSelectLabel } from './select_label';
+import { initSelectLabel, initSelectLabelCategory } from './select_box';
+import { showLabel } from '../components/label';
 
 // Internal imports, e.g:
 // import { initSelect2 } from '../components/init_select2';
@@ -75,6 +76,14 @@ document.addEventListener('turbolinks:load', () => {
 
   if (document.querySelector(".label-choices")) {
     initSelectLabel();
+  }
+
+  if (document.querySelector(".card-label")) {
+    showLabel();
+  }
+
+  if (document.querySelector(".category-choices")) {
+    initSelectLabelCategory();
   }
   
 });
