--- app/javascript/packs/select_label.js
@@ -1,13 +1,0 @@
-const initSelectLabel = () => {
-	let labels = [];
-	labels = document.querySelectorAll(".label-choice")
-	labels.forEach((label) => {
-		label.addEventListener( "click", (event) => {
-			console.log(label);
-			console.log(event.currentTarget);
-			label.classList.toggle("active");
-		});
-	});
-}
-
-export { initSelectLabel };
